# spec/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Emergency Filing Service API",
    "description": "Emergency Filing Service API specification",
    "version": "0.01"
  },
  "basePath": "/efs-submissions-api",
  "schemes": [
    "https"
  ],
...
  "parameters": {
...
    "email-address": {
      "name": "email-address",
      "in": "path",
      "description": "An email address",
      "required": true,
      "type": "string",
      "format": "email",
      "x-example": "demo@ch.gov.uk"
    }
  },
  "responses": {
...
    "BadRequest": {
      "description": "Incomplete or malformed request"
    },
...
    "Unauthorised": {
      "description": "Invalid or missing credentials"
    }
  },
  "paths": {
...
    "/company-authentication/allow-list/{email-address}": {
      "parameters": [
        {
          "$ref": "#/parameters/email-address"
        }
      ],
      "get": {
        "summary": "Check whether an email address is on the allow list",
        "x-operationName": "isOnAllowList",
        "tags": [
          "category",
          "query"
        ],
        "operationId": "isOnAllowList",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Check allow list call is successful",
            "schema": {
              "$ref": "#/definitions/booleanResult"
            }
          },
          "400": {
            "$ref": "#/responses/BadRequest"
          },
          "401": {
            "$ref": "#/responses/Unauthorised"
          }
        }
      }
    }
  },
  "definitions": {
...
    "booleanResult": {
      "type": "boolean",
      "title": "Boolean result",
      "description": "true or false",
      "example": true
    }
  }
}